pyproject.toml
[project]
name = "fishmael"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.10.10",
    "disagain",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
typing = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.9",
]

[tool.uv.sources]
disagain = { git = "https://github.com/sharp-eyes/disagain.git" }


[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "ALL",  # Include everything, including new rules.
]
ignore = [
    "D",  # Ignore docstrings for the time being.
    "D105",  # __magic_method__ docstrings.
    "D107",  # __init__ docstrings.
    "FIX",  # TODO-comments are to-do for a reason. 
    "PLC0414",  # reexports through from x import y as y.
    "PYI041",  # int | float typehints.
    "S101",  # Assert statements; we use these mainly for type assertions.
    "TD",  # TODO-comments are to-do for a reason.
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F403"]
"test.py" = ["ALL"]
